Rename mbe_expander for consistency
